Pairs & Triple in Kotlin -----------------------------------
Pair & Triple is used to store or return two & three values
respectively of same or different data types.
There can or cannot be any relation between both the values.
They can be of the same or different data types.

And how do I use it?
_____________________________

Initialise it as a simple Pair class object by
passing two values to the constructor.
    Ex-
        Pair ("Hello", "How are you")
        Pair ("First", 1)
        Pair (10, null)
        Pair (1.5, listOf(null))
        var variable1 = "I am a String"
        var variable2 = 10

        Pair (variable1, variable2)

And how do I retrieve the values?
______________________________________

Either receive them in a single variable and
use first and second properties or componentN
method to extract the values, or receive in separate
variables and use them directly.

    Ex
        var pair = Pair("Hello", "How are you")
        println(pair.first)
        println(pair.second)

        println(pair.component1())
        println(pair.component2())
        //or
        var (firstName, lastName) = Pair("Hasan", "Khan")
        println(firstName)
        println(lastName)

- We can even use a normal method like toString() on this object.
    EX-
         val pair = Pair("I am a String", listOf(1,2,3))
         print(pair.toString())

- We can even convert a Pair object to a list using toList().
    Ex-
        val pair = Pair("I am a String", 10)
        val list = pair.toList()
        println(list[0])
        println(list.get(1))